Split build into 2 stages
Prior to this change, the whole build was done in a single stage,
including building supercollider, jack and then installing autostart.
If there were any changes that needed to be made to autostart, the whole
build process would require compiling jack and supercollider.

This change installs supercollider and jack and outputs an image file.
This image file can now be used in the second stage to install
autostart. Now, the build process takes around 10 seconds to build the
autostart stage instead of 25 minutes.